refactor - move EventEmitter types to core root
