+++ modules/core/src/index.ts
@@ -31,3 +31,4 @@ export * from './model-params';
 export * from './range';
 export * from './json-ptr';
 export * from './task-loop';
+export * from './events';
